replace rabl template
