+++ app/controllers/projects_controller.rb
@@ -199,7 +199,7 @@ class ProjectsController < ApplicationController
       flash[:error] = I18n.t(:error_can_not_archive_project)
       redirect_back fallback_location: projects_url
     end
-    
+
     update_demo_project_settings @project, false
   end
 
@@ -231,10 +231,10 @@ class ProjectsController < ApplicationController
   end
 
   def level_list
-    @projects = Project.project_level_list(Project.visible)
+    projects = Project.project_level_list(Project.visible)
 
     respond_to do |format|
-      format.api
+      format.json { render json: projects_level_list_json(projects) }
     end
   end
 
@@ -242,6 +242,7 @@ class ProjectsController < ApplicationController
 
   def find_optional_project
     return true unless params[:id]
+
     @project = Project.find(params[:id])
     authorize
   rescue ActiveRecord::RecordNotFound
