+++ spec/features/time_entry/time_entry_report_spec.rb
@@ -34,49 +34,24 @@ describe 'time entry report', type: :feature, js: true do
   let(:work_package) { FactoryBot.create(:work_package, project: project) }
   let!(:project_time_entry) {
     FactoryBot.create_list(:time_entry,
-                            2,
-                            project: project,
-                            work_package: work_package,
-                            hours: 2.5)
+                           2,
+                           project: project,
+                           work_package: work_package,
+                           hours: 2.5)
   }
   let(:project2) { FactoryBot.create(:project) }
   let(:work_package2) { FactoryBot.create(:work_package, project: project2) }
   let!(:project_time_entry2) {
     FactoryBot.create(:time_entry,
-                       project: project2,
-                       spent_on: 1.year.ago,
-                       work_package: work_package2,
-                       hours: 5.0)
+                      project: project2,
+                      spent_on: 1.year.ago,
+                      work_package: work_package2,
+                      hours: 5.0)
   }
   let(:user) { FactoryBot.create(:admin) }
 
   before do
     login_as(user)
-  end
-
-  describe 'details' do
-    context 'for a single project' do
-      before do
-        visit project_time_entries_path(project.identifier)
-      end
-
-      it 'should list the time entries' do
-        expect(page).to have_selector('tr.time-entry', count: 2)
-        expect(page).to have_selector('.time-entry .hours', text: 2.5, count: 2)
-      end
-    end
-
-    context 'for all projects' do
-      before do
-        visit time_entries_path
-      end
-
-      it 'should list the time entries' do
-        expect(page).to have_selector('tr.time-entry', count: 3)
-        expect(page).to have_selector('.time-entry .hours', text: 2.5, count: 2)
-        expect(page).to have_selector('.time-entry .hours', text: 5.0)
-      end
-    end
   end
 
   describe 'reports' do
